Extract aoc leaderboards to dedicated module
